pyproject.toml
[project]
name = "modbus-info"
version = "0.1.0"
description = "从Modbus协议PDF文件中提取关键点位信息并导出为CSV"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.2",
    "openai>=1.54.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.0",
    "mineru[core]>=2.6.8",
    "json-repair>=0.54.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: 集成测试（需要真实API和数据）",
    "slow: 慢速测试",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.black]
line-length = 120
target-version = ["py312"]
